requirements.txt
psycopg2
flask
pyyaml
pyjwt
pillow
hashlib